Correction for raw data location.

+++ code/utility.py
@@ -16,7 +16,7 @@ class Dir_Structure:
 
         # Define paths to directories
         self.par_dir = os.path.dirname(self.code_dir)  # parent directory
-        self.raw_data_dir = os.path.join(self.par_dir, "raw_data", self.model_name)
+        self.raw_data_dir = os.path.join(self.par_dir, "data", "raw_data")
         self.data_dir = os.path.join(self.par_dir, "data", self.model_name)  # stores data/input
         self.output_dir = os.path.join(self.par_dir, "output", self.model_name)  # stores inferrence results
         self.logs_dir = os.path.join(self.par_dir, "logs", self.model_name)  # training log for tensorboard
@@ -28,18 +28,20 @@ class Dir_Structure:
         self.shuffled_indices_path = os.path.join(self.data_dir, "shuffled_indices_{}.npy".format(self.model_name))
         self.input_trainval_path = os.path.join(self.data_dir, "input_trainval.pkl")
         self.output_trainval_path = os.path.join(self.data_dir, "output_trainval.pkl")
-        self.raw_data_path = os.path.join(self.data_dir, "input_values_for_M_by_N_creating_script.csv")
-        self.raw_data_validity_path = os.path.join(self.data_dir, "input_validity_flags_for_M_by_N_creating_script.csv")
+        self.raw_data_path = os.path.join(self.raw_data_dir, "input_values_for_M_by_N_creating_script.csv")
+        self.raw_data_validity_path = os.path.join(self.raw_data_dir, "input_validity_flags_for_M_by_N_creating_script.csv")
         self.pred_trainval_path = os.path.join(self.output_dir, "pred_trainval.pkl")
         self.training_hist_path = os.path.join(self.diag_dir, "training_history.npy")
 
         # clear all contents in the log directory
-        shutil.rmtree(self.logs_dir)
+        if os.path.exists(self.logs_dir):
+            shutil.rmtree(self.logs_dir)
 
         # make these directories if they do not already exist
         self.make_directories()
 
     def make_directories(self):
-        for folder in [self.data_dir, self.output_dir, self.logs_dir, self.ckpts_dir, self.models_dir, self.diag_dir]:
+        for folder in [self.data_dir, self.raw_data_dir, self.output_dir,
+                       self.logs_dir, self.ckpts_dir, self.models_dir, self.diag_dir]:
             if not os.path.exists(folder):
                 os.mkdir(folder)
